requirements.txt
google-cloud-firestore
matplotlib
pandas
pre-commit
seaborn
